+++ modal-go/test/image_test.go
@@ -1,0 +1,39 @@
+package test
+
+import (
+	"context"
+	"testing"
+
+	"github.com/modal-labs/libmodal/modal-go"
+	"github.com/onsi/gomega"
+)
+
+func TestImageFromRegistry(t *testing.T) {
+	t.Parallel()
+	g := gomega.NewWithT(t)
+
+	app, err := modal.AppLookup(context.Background(), "libmodal-test", &modal.LookupOptions{CreateIfMissing: true})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	image, err := app.ImageFromRegistry("alpine:3.21")
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+	g.Expect(image.ImageId).Should(gomega.HavePrefix("im-"))
+}
+
+func TestImageFromAwsEcr(t *testing.T) {
+	t.Parallel()
+	g := gomega.NewWithT(t)
+
+	app, err := modal.AppLookup(context.Background(), "libmodal-test", &modal.LookupOptions{CreateIfMissing: true})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	secret, err := modal.SecretFromName(context.Background(), "aws-ecr-private-registry-test-secret", &modal.SecretFromNameOptions{
+		Environment:  "libmodal",
+		RequiredKeys: []string{"AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY"},
+	})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	image, err := app.ImageFromAwsEcr("459781239556.dkr.ecr.us-east-1.amazonaws.com/ecr-private-registry-test-7522615:python", secret)
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+	g.Expect(image.ImageId).Should(gomega.HavePrefix("im-"))
+}