+++ modal-js/src/config.ts
@@ -10,6 +10,7 @@ interface Config {
     token_id?: string;
     token_secret?: string;
     environment?: string;
+    imageBuilderVersion?: string;
     active?: boolean;
   };
 }
@@ -20,6 +21,7 @@ export interface Profile {
   tokenId: string;
   tokenSecret: string;
   environment?: string;
+  imageBuilderVersion?: string;
 }
 
 async function readConfigFile(): Promise<Config> {
@@ -64,6 +66,9 @@ export function getProfile(profileName?: string): Profile {
     tokenId: process.env["MODAL_TOKEN_ID"] || profileData.token_id,
     tokenSecret: process.env["MODAL_TOKEN_SECRET"] || profileData.token_secret,
     environment: process.env["MODAL_ENVIRONMENT"] || profileData.environment,
+    imageBuilderVersion:
+      process.env["MODAL_IMAGE_BUILDER_VERSION"] ||
+      profileData.imageBuilderVersion,
   };
   if (!profile.tokenId || !profile.tokenSecret) {
     throw new Error(
@@ -77,4 +82,8 @@ export const profile = getProfile(process.env["MODAL_PROFILE"] || undefined);
 
 export function environmentName(environment?: string): string {
   return environment || profile.environment || "";
+}
+
+export function imageBuilderVersion(version?: string): string {
+  return version || profile.imageBuilderVersion || "2024.10";
 }