+++ modal-js/src/image.ts
@@ -1,10 +1,14 @@
+import type {
+  GenericResult,
+  ImageMetadata,
+  ImageRegistryConfig,
+} from "../proto/modal_proto/api";
 import {
   DeploymentNamespace,
-  GenericResult,
   GenericResult_GenericStatus,
-  ImageMetadata,
 } from "../proto/modal_proto/api";
 import { client } from "./client";
+import { imageBuilderVersion } from "./config";
 
 /** A container image, used for starting sandboxes. */
 export class Image {
@@ -19,14 +23,16 @@ export class Image {
 export async function fromRegistryInternal(
   appId: string,
   tag: string,
+  imageRegistryConfig?: ImageRegistryConfig,
 ): Promise<Image> {
   const resp = await client.imageGetOrCreate({
     appId,
     image: {
       dockerfileCommands: [`FROM ${tag}`],
+      imageRegistryConfig: imageRegistryConfig,
     },
     namespace: DeploymentNamespace.DEPLOYMENT_NAMESPACE_WORKSPACE,
-    builderVersion: "2024.10", // TODO: make this configurable
+    builderVersion: imageBuilderVersion(),
   });
 
   let result: GenericResult;
@@ -83,6 +89,5 @@ export async function fromRegistryInternal(
       `Image build for ${resp.imageId} failed with unknown status: ${result.status}`,
     );
   }
-
   return new Image(resp.imageId);
 }