+++ modal-js/src/sandbox.ts
@@ -1,8 +1,7 @@
 import { FileDescriptor } from "../proto/modal_proto/api";
 import { client, isRetryableGrpc } from "./client";
+import type { ModalReadStream, ModalWriteStream } from "./streams";
 import {
-  ModalReadStream,
-  ModalWriteStream,
   streamConsumingIter,
   toModalReadStream,
   toModalWriteStream,